Provide only one static method inside ChatUtil#getAverageColor to eliminate code duplication.
