+++ src/main/java/com/zestarr/pluginportal/commands/subcommands/PreviewSubCommand.java
@@ -9,9 +9,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
-
 public class PreviewSubCommand extends SubCommandManager {
     @Override
     public void execute(CommandSender sender, String[] args, SubCommandEnum subCommandEnum) {
@@ -28,27 +25,5 @@ public class PreviewSubCommand extends SubCommandManager {
 
             previewingPlugin.sendPreview((Player) sender, true);
         });
-    }
-
-    public static Color getAverageColor(BufferedImage bi) {
-        int step = 5;
-
-        int sampled = 0;
-        long sumr = 0, sumg = 0, sumb = 0;
-        for (int x = 0; x < bi.getWidth(); x++) {
-            for (int y = 0; y < bi.getHeight(); y++) {
-                if (x % step == 0 && y % step == 0) {
-                    Color pixel = new Color(bi.getRGB(x, y));
-                    sumr += pixel.getRed();
-                    sumg += pixel.getGreen();
-                    sumb += pixel.getBlue();
-                    sampled++;
-                }
-            }
-        }
-        int dim = bi.getWidth() * bi.getHeight();
-        // Log.info("step=" + step + " sampled " + sampled + " out of " + dim + " pixels (" + String.format("%.1f", (float)(100*sampled/dim)) + " %)");
-        return new Color(Math.round(sumr / sampled), Math.round(sumg / sampled), Math.round(sumb / sampled));
-        // TODO sumr / sampled: Integer division in floating-point context
     }
 }
