+++ src/main/java/com/zestarr/pluginportal/type/PreviewingPlugin.java
@@ -184,7 +184,7 @@ public class PreviewingPlugin {
                     row++;
                 }
                 i++;
-                Color color = getAverageColor(bound);
+                Color color = ChatUtil.getAverageColor(bound);
                 builder.append(ChatColor.of(color)).append("▉");
 
                 switch (i) {
@@ -201,27 +201,5 @@ public class PreviewingPlugin {
             e.printStackTrace();
         }
         player.sendMessage(ChatUtil.format("&8-----------------------------------------------------"));
-    }
-
-    public Color getAverageColor(BufferedImage bi) {
-        int step = 5;
-
-        int sampled = 0;
-        long sumr = 0, sumg = 0, sumb = 0;
-        for (int x = 0; x < bi.getWidth(); x++) {
-            for (int y = 0; y < bi.getHeight(); y++) {
-                if (x % step == 0 && y % step == 0) {
-                    Color pixel = new Color(bi.getRGB(x, y));
-                    sumr += pixel.getRed();
-                    sumg += pixel.getGreen();
-                    sumb += pixel.getBlue();
-                    sampled++;
-                }
-            }
-        }
-        int dim = bi.getWidth() * bi.getHeight();
-        // Log.info("step=" + step + " sampled " + sampled + " out of " + dim + " pixels (" + String.format("%.1f", (float)(100*sampled/dim)) + " %)");
-        return new Color(Math.round(sumr / sampled), Math.round(sumg / sampled), Math.round(sumb / sampled));
-        // TODO sumr / sampled: Integer division in floating-point context
     }
 }
